Add helper file for macro folder
