+++ src/handlers/index.js
@@ -1,4 +1,4 @@
 export { default as handleArbitraryCss } from './arbitraryCss'
-export { default as handleCss } from './css'
+export { default as handleShortCss } from './shortCss'
 export { default as handleDynamic } from './dynamic'
 export { default as handleUserPlugins } from './userPlugins'
